+++ packages/json-rpc-engine/tests/utils.ts
@@ -6,11 +6,8 @@ import type { JsonRpcNotification } from '../src/v2/utils';
 
 const jsonrpc = '2.0' as const;
 
-export const makeRequest = <
-  Input extends Partial<JsonRpcRequest>,
-  Output extends Input & JsonRpcRequest,
->(
-  request: Input = {} as Input,
+export const makeRequest = <Request extends JsonRpcRequest = JsonRpcRequest>(
+  request: Partial<Request> = {},
 ) =>
   ({
     jsonrpc,
@@ -19,7 +16,7 @@ export const makeRequest = <
 
     params: request.params === undefined ? [] : request.params,
     ...request,
-  }) as Output;
+  }) as Request;
 
 export const makeNotification = <Request extends Partial<JsonRpcRequest>>(
   params: Request = {} as Request,