+++ packages/account-tree-controller/src/AccountTreeController.ts
@@ -198,9 +198,8 @@ export class AccountTreeController extends BaseController<
     // Initialize backup and sync config
     this.#backupAndSyncConfig = {
       emitAnalyticsEventFn: (event: BackupAndSyncEmitAnalyticsEventParams) => {
-        return (
-          config?.backupAndSync?.onBackupAndSyncEvent &&
-          config.backupAndSync.onBackupAndSyncEvent(formatAnalyticsEvent(event))
+        return config?.backupAndSync?.onBackupAndSyncEvent?.(
+          formatAnalyticsEvent(event),
         );
       },
     };
@@ -486,7 +485,7 @@ export class AccountTreeController extends BaseController<
 
     for (const id of group.accounts) {
       const account = this.messenger.call('AccountsController:getAccount', id);
-      if (!account || !account.metadata.name.length) {
+      if (!account?.metadata.name.length) {
         continue;
       }
 
@@ -1153,7 +1152,7 @@ export class AccountTreeController extends BaseController<
     const selectedAccount = this.messenger.call(
       'AccountsController:getSelectedMultichainAccount',
     );
-    if (selectedAccount && selectedAccount.id) {
+    if (selectedAccount?.id) {
       const accountMapping = this.#accountIdToContext.get(selectedAccount.id);
       if (accountMapping) {
         const { groupId } = accountMapping;