+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix source network fees for batch Relay deposits on EIP-7702 networks ([#7323](https://github.com/MetaMask/core/pull/7323))
 - Improve Relay provider fees ([#7313](https://github.com/MetaMask/core/pull/7313))
   - Include slippage from feature flag.
   - Read fee from `totalImpact` property.